Factor out `kontrol` module from kevm-pyk codebase (#1985)
* kevm-pyk/src/kontrol: separate module for kontrol

* src/kevm_pyk: remove kontrol related functionality

* kevm-pyk/pyproject: expose kontrol binary

* kevm-pyk/src/kontrol/__init__: add stub init module

* kevm_pyk/src/tests: adjust for new kontrol module

* bin/kevm: route foundry-* commands through kontrol module

* Set Version: 1.0.244

* flake.nix: patch kontrol calls for nix builds too

* bin/kevm: add kontrol help to help message

* flake.nix: fix flake

* bin/kevm: correct syntax error

* bin/kevm, flake.nix: simplify calls to gst-to-kore/kontrol

* bin/kevm, flake.nix: simplify how kevm-pyk is invoked

* {kontrol,kevm_pyk}/__main__: adjust CLI name

* kevm-pyk/pyproject.toml: publish two separate packages

* package/docker/Dockerfile: try setting up PATH appropriately

* package/docker/Dockerfile: use newer version of poetry

* package/docker/Dockerfile: correct PATH

* Set Version: 1.0.250

* kevm-pyk/pyproject: publish packages correctly

* Set Version: 1.0.251

* Set Version: 1.0.252

* package/docker/Dockerfile: correct PATH

* Set Version: 1.0.254

* flake.nix: correct substitution for new file location

* Set Version: 1.0.261

* .github/test-pr: correct to normal runner for booster tests

---------
