+++ .github/workflows/test-pr.yml
@@ -171,7 +171,7 @@ jobs:
   test-prove-foundry-booster:
     name: 'Build and Test KEVM Foundry proofs (booster)'
     needs: kevm-pyk-code-quality-checks
-    runs-on: [self-hosted, linux, huge]
+    runs-on: [self-hosted, linux, normal]
     timeout-minutes: 150
     steps:
       - name: 'Check out code'
