+++ flake.nix
@@ -91,9 +91,11 @@
               substituteInPlace ./cmake/node/CMakeLists.txt \
                 --replace 'set(K_LIB ''${K_BIN}/../lib)' 'set(K_LIB ${k}/lib)'
               substituteInPlace ./bin/kevm \
-                --replace 'execute python3 -m kevm_pyk' 'execute ${final.kevm-pyk}/bin/kevm-pyk'
+                --replace 'execute kevm-pyk' 'execute ${final.kevm-pyk}/bin/kevm-pyk'
               substituteInPlace ./bin/kevm \
-                --replace 'gst-to-kore' '${final.kevm-pyk}/bin/gst-to-kore'
+                --replace 'execute kontrol' 'execute ${final.kevm-pyk}/bin/kontrol'
+              substituteInPlace ./bin/kevm \
+                --replace 'execute gst-to-kore' 'execute ${final.kevm-pyk}/bin/gst-to-kore'
             '';
 
             buildFlagsArray = "NIX_LIBS=${nixLibs prev}";
@@ -154,7 +156,7 @@
           projectDir = ./kevm-pyk;
 
           postPatch = ''
-              substituteInPlace ./src/kevm_pyk/foundry.py \
+              substituteInPlace ./src/kontrol/foundry.py \
                 --replace "'forge', 'build'," "'forge', 'build', '--no-auto-detect',"
             '';
