+++ kevm-pyk/pyproject.toml
@@ -4,10 +4,14 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "kevm-pyk"
-version = "1.0.260"
+version = "1.0.261"
 description = ""
 authors = [
     "Runtime Verification, Inc. <contact@runtimeverification.com>",
+]
+packages = [
+    { include = "kevm_pyk", from = "src/" },
+    { include = "kontrol", from = "src/" },
 ]
 
 [tool.poetry.dependencies]
@@ -36,6 +40,7 @@ pyupgrade = "*"
 [tool.poetry.scripts]
 kevm-pyk = "kevm_pyk.__main__:main"
 gst-to-kore = "kevm_pyk.gst_to_kore:main"
+kontrol = "kontrol.__main__:main"
 
 [tool.isort]
 profile = "black"
