+++ kevm-pyk/src/tests/integration/test_foundry_prove.py
@@ -8,7 +8,7 @@ from filelock import FileLock
 from pyk.utils import run_process
 
 from kevm_pyk import config
-from kevm_pyk.foundry import foundry_kompile, foundry_prove, foundry_show
+from kontrol.foundry import foundry_kompile, foundry_prove, foundry_show
 
 from .utils import TEST_DATA_DIR
